@@ -12,16 +12,15 @@ classifiers =
     Programming Language :: Python :: 3.10
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.9
 packages = find:
 package_dir =
     =src
 install_requires =
     bluesky
     pyepics
-    blueapi @ git+https://github.com/DiamondLightSource/blueapi@main
+    blueapi
     flask-restful
-    zocalo
     ispyb
     scanspec
     numpy
@@ -36,8 +35,9 @@ install_requires =
     xarray
     doct
     databroker
-    dls-dodal @ git+https://github.com/DiamondLightSource/dodal.git@2c40b539c9c92e6f6c14ff609158aa3969f1e515
+    dls-dodal @ git+https://github.com/DiamondLightSource/dodal.git@12bbf4cd438456186f8ca694789b20b0708063f8
     pydantic<2.0 # See https://github.com/DiamondLightSource/hyperion/issues/774
+    scipy
 
 
 [options.extras_require]
@@ -46,6 +46,7 @@ dev =
     black
     pytest-cov
     pytest-random-order
+    pytest-asyncio
     ipython
     mockito
     pre-commit
@@ -54,6 +55,10 @@ dev =
     tox
     build
     ruff
+    diff-cover
+    pyright
+    pyright_diff_quality_plugin @ git+https://github.com/DiamondLightSource/pyright_diff_quality_plugin.git
+
 
 [options.packages.find]
 where = src
