@@ -5,7 +5,7 @@ from subprocess import PIPE, CalledProcessError, Popen
 from git import Repo
 from packaging.version import Version
 
-recognised_beamlines = ["dev", "i03"]
+recognised_beamlines = ["dev", "i03", "i04"]
 
 
 class repo:
@@ -67,10 +67,8 @@ def get_hyperion_release_dir_from_args(repo: repo) -> str:
     if args.beamline == "dev":
         print("Running as dev")
         return "/tmp/hyperion_release_test/bluesky"
-    elif args.beamline == "i03":
-        return f"/dls_sw/{args.beamline}/software/bluesky"
     else:
-        raise Exception("not running in dev mode, exiting... (remove this)")
+        return f"/dls_sw/{args.beamline}/software/bluesky"
 
 
 if __name__ == "__main__":
