@@ -16,7 +16,7 @@ from dodal.devices.eiger import EigerDetector
 from dodal.devices.fast_grid_scan import FastGridScan
 from dodal.devices.flux import Flux
 from dodal.devices.oav.oav_detector import OAV
-from dodal.devices.oav.oav_parameters import OAV_CONFIG_FILE_DEFAULTS, OAVParameters
+from dodal.devices.oav.oav_parameters import OAV_CONFIG_JSON, OAVParameters
 from dodal.devices.oav.pin_image_recognition import PinTipDetection
 from dodal.devices.s4_slit_gaps import S4SlitGaps
 from dodal.devices.smargon import Smargon
@@ -24,6 +24,7 @@ from dodal.devices.synchrotron import Synchrotron
 from dodal.devices.undulator import Undulator
 from dodal.devices.xbpm_feedback import XBPMFeedback
 from dodal.devices.zebra import Zebra
+from dodal.devices.zocalo import ZocaloResults
 
 from hyperion.device_setup_plans.utils import (
     start_preparing_data_collection_then_do_plan,
@@ -78,6 +79,7 @@ class GridDetectThenXRayCentreComposite:
     undulator: Undulator
     xbpm_feedback: XBPMFeedback
     zebra: Zebra
+    zocalo: ZocaloResults
 
     def __post_init__(self):
         """Ensure that aperture positions are loaded whenever this class is created."""
@@ -88,21 +90,6 @@ class GridDetectThenXRayCentreComposite:
 
 def create_devices(context: BlueskyContext) -> GridDetectThenXRayCentreComposite:
     return device_composite_from_context(context, GridDetectThenXRayCentreComposite)
-
-
-def wait_for_det_to_finish_moving(detector: DetectorMotion, timeout=120.0):
-    LOGGER.info("Waiting for detector to finish moving")
-    SLEEP_PER_CHECK = 0.1
-    times_to_check = int(timeout / SLEEP_PER_CHECK)
-    for _ in range(times_to_check):
-        detector_state = yield from bps.rd(detector.shutter)
-        detector_z_dmov = yield from bps.rd(detector.z.motor_done_move)
-        LOGGER.info(f"Shutter state is {'open' if detector_state==1 else 'closed'}")
-        LOGGER.info(f"Detector z DMOV is {detector_z_dmov}")
-        if detector_state == 1 and detector_z_dmov == 1:
-            return
-        yield from bps.sleep(SLEEP_PER_CHECK)
-    raise TimeoutError("Detector not finished moving")
 
 
 def create_parameters_for_flyscan_xray_centre(
@@ -131,7 +118,9 @@ def detect_grid_and_do_gridscan(
 
     oav_callback = OavSnapshotCallback()
     grid_params_callback = GridDetectionCallback(
-        oav_params, experiment_params.exposure_time
+        composite.oav.parameters,
+        experiment_params.exposure_time,
+        experiment_params.set_stub_offsets,
     )
 
     @bpp.subs_decorator([oav_callback, grid_params_callback])
@@ -190,7 +179,6 @@ def detect_grid_and_do_gridscan(
         composite.aperture_scatterguard,
         composite.aperture_scatterguard.aperture_positions.SMALL,
     )
-    yield from wait_for_det_to_finish_moving(composite.detector_motion)
 
     flyscan_composite = FlyScanXRayCentreComposite(
         aperture_scatterguard=composite.aperture_scatterguard,
@@ -205,6 +193,7 @@ def detect_grid_and_do_gridscan(
         synchrotron=composite.synchrotron,
         xbpm_feedback=composite.xbpm_feedback,
         zebra=composite.zebra,
+        zocalo=composite.zocalo,
     )
 
     yield from flyscan_xray_centre(
@@ -216,7 +205,7 @@ def detect_grid_and_do_gridscan(
 def grid_detect_then_xray_centre(
     composite: GridDetectThenXRayCentreComposite,
     parameters: Any,
-    oav_param_files: dict = OAV_CONFIG_FILE_DEFAULTS,
+    oav_config: str = OAV_CONFIG_JSON,
 ) -> MsgGenerator:
     """
     A plan which combines the collection of snapshots from the OAV and the determination
@@ -226,7 +215,7 @@ def grid_detect_then_xray_centre(
 
     eiger.set_detector_parameters(parameters.hyperion_params.detector_params)
 
-    oav_params = OAVParameters("xrayCentring", **oav_param_files)
+    oav_params = OAVParameters("xrayCentring", oav_config)
 
     plan_to_perform = detect_grid_and_do_gridscan(
         composite,
@@ -236,5 +225,7 @@ def grid_detect_then_xray_centre(
 
     return start_preparing_data_collection_then_do_plan(
         eiger,
+        composite.detector_motion,
+        parameters.hyperion_params.detector_params.detector_distance,
         plan_to_perform,
     )
