@@ -9,13 +9,19 @@ import bluesky.plans as bluesky_plans
 from blueapi.core import BlueskyContext, MsgGenerator
 from bluesky.run_engine import RunEngine
 from bluesky.utils import ProgressBarManager
+from dodal.beamlines.beamline_parameters import (
+    BEAMLINE_PARAMETER_PATHS,
+    GDABeamlineParameters,
+)
 from dodal.devices.smargon import Smargon
 from dodal.utils import get_beamline_name
 
 from hyperion.log import LOGGER
 from hyperion.parameters import external_parameters
-from hyperion.parameters.beamline_parameters import GDABeamlineParameters
-from hyperion.parameters.constants import BEAMLINE_PARAMETER_PATHS, SIM_BEAMLINE
+from hyperion.parameters.constants import (
+    GRIDSCAN_MAIN_PLAN,
+    SIM_BEAMLINE,
+)
 from hyperion.tracing import TRACER
 from hyperion.utils.context import device_composite_from_context, setup_context
 
@@ -47,7 +53,7 @@ def run_gridscan(
     composite: SteppedGridScanComposite,
     parameters: SteppedGridScanInternalParameters,
     md={
-        "plan_name": "run_gridscan",
+        "plan_name": GRIDSCAN_MAIN_PLAN,
     },
 ):
     sample_motors: Smargon = composite.smargon
