@@ -1,16 +1,23 @@
 import numpy as np
 from bluesky.callbacks import CallbackBase
 from dodal.devices.fast_grid_scan import GridScanParams
-from dodal.devices.oav.oav_parameters import OAVParameters
+from dodal.devices.oav.oav_detector import OAVConfigParams
 
 from hyperion.device_setup_plans.setup_oav import calculate_x_y_z_of_pixel
 from hyperion.log import LOGGER
 
 
 class GridDetectionCallback(CallbackBase):
-    def __init__(self, oav_params: OAVParameters, exposure_time: float, *args) -> None:
+    def __init__(
+        self,
+        oav_params: OAVConfigParams,
+        exposure_time: float,
+        set_stub_offsets: bool,
+        *args,
+    ) -> None:
         super().__init__(*args)
         self.exposure_time = exposure_time
+        self.set_stub_offsets = set_stub_offsets
         self.oav_params = oav_params
         self.start_positions: list = []
         self.box_numbers: list = []
@@ -51,7 +58,7 @@ class GridDetectionCallback(CallbackBase):
 
     def get_grid_parameters(self) -> GridScanParams:
         return GridScanParams(
-            dwell_time=self.exposure_time * 1000,
+            dwell_time_ms=self.exposure_time * 1000,
             x_start=self.start_positions[0][0],
             y1_start=self.start_positions[0][1],
             y2_start=self.start_positions[0][1],
@@ -63,4 +70,5 @@ class GridDetectionCallback(CallbackBase):
             x_step_size=self.x_step_size_mm,
             y_step_size=self.y_step_size_mm,
             z_step_size=self.z_step_size_mm,
+            set_stub_offsets=self.set_stub_offsets,
         )
