@@ -1,15 +1,17 @@
 from __future__ import annotations
 
-from bluesky.callbacks import CallbackBase
-
+from hyperion.external_interaction.callbacks.plan_reactive_callback import (
+    PlanReactiveCallback,
+)
 from hyperion.external_interaction.nexus.write_nexus import NexusWriter
-from hyperion.log import LOGGER
+from hyperion.log import NEXUS_LOGGER
+from hyperion.parameters.constants import ROTATION_OUTER_PLAN
 from hyperion.parameters.plan_specific.rotation_scan_internal_params import (
     RotationInternalParameters,
 )
 
 
-class RotationNexusFileCallback(CallbackBase):
+class RotationNexusFileCallback(PlanReactiveCallback):
     """Callback class to handle the creation of Nexus files based on experiment
     parameters for rotation scans
 
@@ -24,20 +26,21 @@ class RotationNexusFileCallback(CallbackBase):
     Usually used as part of a RotationCallbackCollection.
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
+        super().__init__()
         self.run_uid: str | None = None
         self.parameters: RotationInternalParameters | None = None
         self.writer: NexusWriter | None = None
 
-    def start(self, doc: dict):
-        if doc.get("subplan_name") == "rotation_scan_with_cleanup":
+    def activity_gated_start(self, doc: dict):
+        if doc.get("subplan_name") == ROTATION_OUTER_PLAN:
             self.run_uid = doc.get("uid")
-            LOGGER.info(
+            NEXUS_LOGGER.info(
                 "Nexus writer recieved start document with experiment parameters."
             )
             json_params = doc.get("hyperion_internal_parameters")
             self.parameters = RotationInternalParameters.from_json(json_params)
-            LOGGER.info("Setting up nexus file.")
+            NEXUS_LOGGER.info("Setting up nexus file.")
             self.writer = NexusWriter(
                 self.parameters,
                 self.parameters.get_scan_points(),
