@@ -32,10 +32,15 @@ class PinCentreThenXrayCentreParams(AbstractExperimentParameterBase):
     omega_start: float
 
     tip_offset_microns: float = 0
-    oav_centring_file: str = "/dls_sw/i03/software/gda/configurations/i03-config/etc/OAVCentring_hyperion.json"
+    oav_centring_file: str = (
+        "/dls_sw/i03/software/daq_configuration/json/OAVCentring_hyperion.json"
+    )
 
     # Width for single pin
     grid_width_microns: float = 600
+
+    # Whether to set the stub offsets after centering
+    set_stub_offsets: bool = False
 
     def get_num_images(self):
         return 0
