@@ -32,7 +32,9 @@ def find_device_in_context(
 
     device = context.find_device(name)
     if device is None:
-        raise ValueError(f"Cannot find device named '{name}' in bluesky context.")
+        raise ValueError(
+            f"Cannot find device named '{name}' in bluesky context {context.devices}."
+        )
 
     if not isinstance(device, expected_type):
         raise ValueError(
@@ -83,5 +85,7 @@ def setup_context(
         wait_for_connection=wait_for_connection,
         fake_with_ophyd_sim=fake_with_ophyd_sim,
     )
+
+    LOGGER.info(f"Plans found in context: {context.plan_functions.keys()}")
 
     return context
